add html string, refactoring, updating
the following commit changed the following: instead having a seperate file parse it and pull information, instead a refactored our code to grab it from the script we use to combine both json and the shapefile. additionally the code will output 2 things: it will output the merged file as the same name (it just has the html string) and it outputs a second file (called *counties/tracts*_poup.json for ease of reading. I also updated the makefile and cleaned up the redundant files no longer needed after this major refactor